% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arg_small}
\alias{arg_small}
\title{Internal utility function.}
\usage{
arg_small(args)
}
\arguments{
\item{args}{named list of arguments to \code{\link{downsize}}}
}
\description{
Utility function. Make downsized object out of  
argument \code{small} to \code{\link{downsize}}.
}
\seealso{
\code{\link{help_downsize}}
}

