% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_args}
\alias{check_args}
\title{Internal utility function.}
\usage{
check_args(args)
}
\arguments{
\item{args}{named list of arguments to \code{\link{downsize}}}
}
\description{
Utility function. Checks that arguments are valid.
}
\seealso{
\code{\link{help_downsize}}
}

