% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_downsized}
\alias{check_downsized}
\title{Internal utility function.}
\usage{
check_downsized(big, small)
}
\arguments{
\item{big}{argument to \code{\link{downsize}}}

\item{small}{argument to \code{\link{downsize}}}
}
\description{
Utility function. Checks if downsizing really happened.
}
\seealso{
\code{\link{downsize}}
}

