% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_small}
\alias{check_small}
\title{Internal utility function.}
\usage{
check_small(big, small)
}
\arguments{
\item{big}{argument to \code{\link{downsize}}}

\item{small}{argument to \code{\link{downsize}}}
}
\description{
Utility function. 
Called if "downsize" is TRUE. Compare "small" to "big" and see if
the result makes sense.
}
\seealso{
\code{\link{downsize}}
}

