% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{downsize}
\alias{downsize}
\title{Function \code{downsize}}
\usage{
downsize()
}
\description{
Same as the \code{\link{scale_down}} function. 
Calls \code{options(downsize = TRUE)} to scale down a workflow.
This affects the \code{ds()} function.
}
\seealso{
\code{\link{ds}}, \code{\link{scale_down}}, \code{\link{scale_up}},
\code{\link{scaling}}
}

