% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsize_error}
\alias{downsize_error}
\title{Internal utility function.}
\usage{
downsize_error(arg_name)
}
\arguments{
\item{arg_name}{name of an argument to \code{\link{downsize}}}
}
\description{
Utility function. Throw an error when downsizing fails.
}
\seealso{
\code{\link{downsize}}
}

