% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{downsize_help}
\alias{downsize_help}
\title{Function \code{downsize_help}}
\usage{
downsize_help()
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\seealso{
\code{\link{downsize}}, \code{\link{scale_down}}, 
\code{\link{scale_up}}, \code{\link{scaling}}
}

