% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Function \code{ds}}
\usage{
ds(big, small = NULL, downsize = downsize, length = NULL, dim = NULL,
  nrow = NULL, ncol = NULL, random = FALSE)
}
\arguments{
\item{big}{Object to be downsized if \code{downsize} is \code{TRUE}.}

\item{small}{Object to return if \code{downsize} is \code{TRUE}. If unspecified,
the return object will be computed from \code{big}.}

\item{downsize}{TRUE/FALSE value, whether to downsize. Defaults to the 
global option \code{downsize}, which you can see with \code{getOption("downsize")}
and set with \code{options(downsize = TRUE)} or \code{options(downsize = FALSE)}.}

\item{length}{Length to downsize vectors to.}

\item{dim}{Dimension to downsize arrays and matrices to.}

\item{nrow}{Downsized number of rows. Overrides \code{dim}.}

\item{ncol}{Downsized number of columns. Overrides \code{dim}.}

\item{random}{If TRUE, take a random subset of \code{big} instead
of the first few elements, rows, columns, etc.}
}
\value{
\code{big} if \code{downsize} is \code{FALSE} and 
a downsized object otherwise (either \code{small} if specified or 
a select few elements of \code{big}).
}
\description{
Downsize an R object if \code{downsize} is \code{TRUE}.
}

