% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_array}
\alias{ds_array}
\title{Utility function}
\usage{
ds_array(big, dim = NULL, random = FALSE)
}
\arguments{
\item{big}{utility parameter}

\item{dim}{utility parameter}

\item{random}{utility parameter}
}
\value{
utility object
}
\description{
Utility function
}
\seealso{
ds
}

