% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_error}
\alias{ds_error}
\title{Internal utility function.}
\usage{
ds_error(arg_name)
}
\arguments{
\item{arg_name}{name of an argument to \code{\link{ds}}}
}
\description{
Utility function. Throw an error when downsizing fails.
}
\seealso{
\code{\link{ds}}
}

