% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_options}
\alias{ds_options}
\title{Internal utility function.}
\usage{
ds_options()
}
\description{
Set the \code{downsize} option to \code{FALSE}. 
Called by \code{.onLoad()}.
}
\seealso{
\code{\link{ds}}
}

