% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ds_vector}
\alias{ds_vector}
\title{Utility function}
\usage{
ds_vector(big, length = NULL, random = FALSE)
}
\arguments{
\item{big}{utility parameter}

\item{length}{utility parameter}

\item{random}{utility parameter}
}
\value{
utility object
}
\description{
Utility function
}
\seealso{
ds
}

