% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_dim}
\alias{fix_dim}
\title{Utility function}
\usage{
fix_dim(big, dim = NULL, length = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{big}{utility parameter}

\item{dim}{utility parameter}

\item{length}{utility parameter}

\item{nrow}{utility parameter}

\item{ncol}{utility parameter}
}
\value{
utility object
}
\description{
Utility function
}
\seealso{
ds
}

