% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_indices}
\alias{get_indices}
\title{Utility function}
\usage{
get_indices(max_length, length, random = FALSE)
}
\arguments{
\item{max_length}{utility parameter}

\item{length}{utility parameter}

\item{random}{utility parameter}
}
\value{
utility object
}
\description{
Utility function
}
\seealso{
ds
}

