% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{help_downsize}
\alias{help_downsize}
\title{Function \code{help_downsize}}
\usage{
help_downsize()
}
\description{
Prints links for tutorials, troubleshooting, bug reports, etc.
}
\seealso{
\code{\link{downsize}}, \code{\link{scale_down}}, 
\code{\link{scale_up}}, \code{\link{scaling}}
}
