% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_small}
\alias{make_small}
\title{Internal utility function.}
\usage{
make_small(args)
}
\arguments{
\item{args}{named list of arguments to \code{\link{downsize}}}
}
\description{
Utility function. Inside function \code{\link{downsize}}, 
make "small" out of "big" 
by subsetting or some other method. This is called if any of 
the subsetting arguments to \code{\link{downsize}} are set (length, dim, etc.)
}
\seealso{
\code{\link{help_downsize}}
}

