% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{production_mode}
\alias{production_mode}
\title{Function \code{production_mode}}
\usage{
production_mode()
}
\description{
Calls \code{options(downsize = FALSE)} to scale up a workflow
to production mode.
This affects the \code{\link{downsize}} function.
Use the \code{\link{help_downsize}} function to get more help.
}
\details{
Use the \code{\link{help_downsize}} function to get more help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{downsize}}, 
\code{\link{test_mode}}, \code{\link{my_mode}}
}

