% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{scale_up}
\alias{scale_up}
\title{Function \code{scale_up}}
\usage{
scale_up()
}
\description{
Call \code{options(downsize = FALSE)} to scale up a workflow.
This affects the \code{\link{downsize}} function.
Use the \code{\link{help_downsize}} function to get more help.
}
\details{
Use the \code{\link{help_downsize}} function to get more help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{downsize}}, 
\code{\link{scale_down}}, \code{\link{scaling}}
}

