% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{scaling}
\alias{scaling}
\title{Function \code{scaling}}
\usage{
scaling()
}
\description{
Check whether the current workflow is scaled up or down.
}
\seealso{
\code{\link{ds}}, \code{\link{scale_down}}, \code{\link{scale_up}}
}

