% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_downsize}
\alias{should_downsize}
\title{Internal utility function.}
\usage{
should_downsize(downsize)
}
\arguments{
\item{downsize}{argument to \code{\link{ds}}}
}
\description{
Utility function. TRUE/FALSE: should the ds() function return 
a downsized replacement for "big"?
}
\seealso{
\code{\link{ds}}
}

