% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{should_make_small}
\alias{should_make_small}
\title{Internal utility function.}
\usage{
should_make_small(args)
}
\arguments{
\item{args}{named list of arguments to \code{\link{ds}}}
}
\description{
Utility function. Should "small" be made from "big"? 
Returns \code{TRUE} if at least one of the subsetting arguments
(length, dim, etc.) or similar is set.
}
\seealso{
\code{\link{ds}}
}

