% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_dim}
\alias{subset_dim}
\title{Internal utility function.}
\usage{
subset_dim(x, dim, random)
}
\arguments{
\item{x}{object to subset}

\item{dim}{new dimensions}

\item{random}{logical, whether to take a random subset or just the head.}
}
\description{
Utility function. Subset an object \code{x} with a 
\code{dim} attribute (matrix/array/data.frame/etc.)
and return another with dimensions pmin(dim, dim(x))
}
\seealso{
\code{\link{help_downsize}}
}

