% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_nrow}
\alias{subset_nrow}
\title{Internal utility function.}
\usage{
subset_nrow(x, nrow, random)
}
\arguments{
\item{x}{object to subset}

\item{nrow}{new number of rows}

\item{random}{logical, whether to take a random subset or just the head.}
}
\description{
Utility function. Subset an object \code{x} with rows and 
return another with min(nrow, nrow(x)) columns.
}
\seealso{
\code{\link{help_downsize}}
}
