% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{test_mode}
\alias{test_mode}
\title{Function \code{test_mode}}
\usage{
test_mode()
}
\description{
Calls \code{options(downsize = TRUE)} to scale down a workflow
to test mode.
This affects the \code{\link{downsize}} function.
Use the \code{\link{help_downsize}} function to get more help.
}
\details{
Use the \code{\link{help_downsize}} function to get more help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{downsize}}, \code{\link{production_mode}}, \code{\link{my_mode}},
\code{\link{my_mode}}
}
