% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_arg_small}
\alias{use_arg_small}
\title{Internal utility function.}
\usage{
use_arg_small(args)
}
\arguments{
\item{args}{named list of arguments to \code{\link{downsize}}}
}
\description{
Utility function. Should the "small" argument to 
\code{\link{downsize}} be used? Returns \code{TRUE} if none 
of the subsetting arguments (length, dim, etc.) or similar is set.
}
\seealso{
\code{\link{help_downsize}}
}
