% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Function \code{ds}}
\usage{
ds(big, small = NULL, downsize = getOption("downsize"), length = NULL,
  dim = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{big}{Big version of the object.}

\item{small}{Small version of the object for a downsized analysis}

\item{downsize}{TRUE/FALSE value, whether to downsize.}

\item{length}{Length to downsize vectors to.}

\item{dim}{Dimension to downsize arrays and matrices to.}

\item{nrow}{Downsized number of rows. Overrides \code{dim}.}

\item{ncol}{Downsized number of columns. Overrides \code{dim}.}
}
\value{
\code{big} if \code{downsize} is \code{TRUE} and \code{small} otherwise.
}
\description{
Conditionally downsize an R object
}

