% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ds}
\alias{ds}
\title{Function \code{ds}. Deprecated.}
\usage{
ds(big, small = NULL, downsize = getOption("downsize"), warn = TRUE,
  random = FALSE, length = NULL, dim = NULL, ncol = NULL, nrow = NULL)
}
\arguments{
\item{big}{Object to return if \code{downsize} is \code{FALSE} or \code{NULL}.}

\item{small}{Object to return if \code{downsize} is \code{TRUE} and all subsetter 
arguments such as \code{length} and \code{dim} are \code{NULL}.}

\item{downsize}{\code{TRUE}/\code{FALSE} value (\code{NULL} counts as \code{FALSE}), 
whether to replace \code{big} with a downsized object.
Defaults to the global option \code{downsize}, which you can check with
\code{getOption("downsize")} or the \code{\link{scaling}} function and set with functions
\code{\link{downsize}}, \code{\link{scale_down}} or \code{\link{scale_up}}.}

\item{warn}{\code{TRUE/FALSE} option to warn the user if \code{big} and \code{small}
are identical or \code{big} is smaller in memory than \code{small}.}

\item{random}{If \code{TRUE}, take a random subset of \code{big} instead
of the first few elements. For example, if \code{nrow == 3}, take a random 
3 rows instead of the first 3.}

\item{length}{Downsize \code{big} to this length if \code{downsize} is \code{TRUE}.}

\item{dim}{Downsize \code{big} to these dimensions if \code{downsize} is \code{TRUE}.}

\item{ncol}{Downsize \code{big} to this number of columns if \code{downsize} is \code{TRUE}.}

\item{nrow}{Downsize \code{big} to this number of rows if \code{downsize} is \code{TRUE}.}
}
\value{
A downsized object if \code{downsize} is \code{TRUE} and \code{big} otherwise.
}
\description{
Deprecated. Use function \code{\link{downsize}} instead.
See \code{\link{help_downsize}} for help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{downsize}}, \code{\link{scale_down}}, 
\code{\link{scale_up}}, \code{\link{scaling}}
}

