% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{scale_down}
\alias{scale_down}
\title{Function \code{scale_down}}
\usage{
scale_down()
}
\description{
Calls \code{options(downsize = TRUE)} to scale down a workflow.
This affects the \code{\link{downsize}} function.
Use the \code{\link{help_downsize}} function to get more help.
}
\details{
Use the \code{\link{help_downsize}} function to get more help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{downsize}}, \code{\link{scale_up}}, \code{\link{scaling}},
\code{\link{scaling}}
}

