% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{scale_down}
\alias{scale_down}
\title{Function \code{scale_down}}
\usage{
scale_down()
}
\description{
Call \code{options(downsize = TRUE)} to scale down a workflow.
This affects the \code{ds()} function.
}
\seealso{
\code{\link{ds}}, \code{\link{scale_up}}, \code{\link{scaling}}
}

