% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{scale_up}
\alias{scale_up}
\title{Function \code{scale_up}}
\usage{
scale_up()
}
\description{
Call \code{options(downsize = TRUE)} to scale up a workflow.
This affects the \code{ds()} function.
}
\seealso{
\code{\link{scale_down}}, \code{\link{ds}}
}

