% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{scale_up}
\alias{scale_up}
\title{Deprecated function \code{scale_up}}
\usage{
scale_up()
}
\description{
Deprecated. Use \code{\link{production_mode}} instead.
Calls \code{options(downsize = FALSE)} to scale up a workflow.
This affects the \code{\link{downsize}} function.
Use the \code{\link{help_downsize}} function to get more help.
}
\details{
Use the \code{\link{help_downsize}} function to get more help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{production_mode}}
}

