% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{scaling}
\alias{scaling}
\title{Deprecated function \code{scaling}}
\usage{
scaling()
}
\value{
\code{"scaled down"} if \code{getOption("downsize")} is \code{TRUE} 
and "scaled up" if \code{getOption("downsize")} is \code{FALSE} or \code{NULL}.
}
\description{
Deprecated. Use \code{\link{my_mode}} instead.
Checks whether the current workflow is scaled up or down.
Use the \code{\link{help_downsize}} function to get more help.
}
\details{
Use the \code{\link{help_downsize}} function to get more help.
}
\seealso{
\code{\link{help_downsize}}, \code{\link{my_mode}}
}

