% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{scaling}
\alias{scaling}
\title{Function \code{scaling}}
\usage{
scaling()
}
\value{
\code{"scaled down"} if \code{getOption("downsize")} is \code{TRUE} 
and "scaled up" if \code{getOption("downsize")} is \code{FALSE} or \code{NULL}.
}
\description{
Check whether the current workflow is scaled up or down.
}
\seealso{
\code{\link{downsize}}, \code{\link{scale_down}}, 
\code{\link{scale_up}}
}

